created base.html as standard and index.html as mainPage. Created main.css as stylesheet and linked with base
