+++ templates/index.html
@@ -1,0 +1,9 @@
+{% extends 'base.html' %}
+
+{% block head %}
+<title>Fratelli - Cucina Italiana Bar</title>
+{% endblock %}
+
+{% block body %}
+<h1>Fratelli - Cucina Italiana Bar</h1>
+{% endblock %}
